requirements.txt
azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.22.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
distlib==0.3.8
filelock==3.15.4
idna==3.7
isodate==0.6.1
logzero==1.7.0
numpy==2.0.1
pandas==2.2.2
pipenv==2024.0.1
platformdirs==4.2.2
pycparser==2.22
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==72.2.0
simplejson==3.19.3
six==1.16.0
smartapi-python==1.4.9
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
